% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{calcul_groupe}
\alias{calcul_groupe}
\title{Grouped calculation: average revenue by group}
\usage{
calcul_groupe(data, groupe)
}
\arguments{
\item{data}{The retail dataset}

\item{groupe}{Grouping variable (eg: "Country")}
}
\value{
A data.frame with average revenue per group
}
\description{
Grouped calculation: average revenue by group
}
