% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{filtrer_pays}
\alias{filtrer_pays}
\title{Filter data by country}
\usage{
filtrer_pays(data, pays)
}
\arguments{
\item{data}{The retail dataset}

\item{pays}{Country name to filter (character)}
}
\value{
A filtered data.frame
}
\description{
Filter data by country
}
