% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{graphique_ca}
\alias{graphique_ca}
\title{ggplot2 chart of average revenue by group}
\usage{
graphique_ca(data, n = 10)
}
\arguments{
\item{data}{Result of `calcul_groupe()`}

\item{n}{Number of groups to display (default: 10)}
}
\value{
A ggplot2 chart
}
\description{
ggplot2 chart of average revenue by group
}
