% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retail}
\alias{retail}
\title{Transactions de vente au détail - Décembre 2010}
\format{
Un data.frame avec 5 variables :
\describe{
  \item{Country}{Pays du client (character)}
  \item{Description}{Nom du produit (character)}
  \item{Quantity}{Quantité achetée (integer)}
  \item{UnitPrice}{Prix unitaire en livres sterling (double)}
  \item{InvoiceDate}{Date et heure de la transaction (POSIXct)}
}
}
\source{
\url{https://www.kaggle.com/datasets/vijayuv/onlineretail}
}
\usage{
retail
}
\description{
Extrait du jeu de données Online Retail UCI, filtré sur décembre 2010.
}
\keyword{datasets}
